+++ src/main/java/site/courseregistrationsystem/enrollment/presentation/EnrollmentController.java
@@ -30,7 +30,8 @@ public class EnrollmentController {
 
 	@ResponseStatus(HttpStatus.CREATED)
 	@PostMapping("/fast/{lectureNumber}")
-	public ApiResponse<EnrolledLecture> enrollLectureByNumber(@Login Long studentPk, @PathVariable Long lectureNumber) {
+	public ApiResponse<EnrolledLecture> enrollLectureByNumber(@Login Long studentPk,
+		@PathVariable Integer lectureNumber) {
 		return ApiResponse.of(HttpStatus.CREATED, ResponseMessage.ENROLL_LECTURE_SUCCESS.getMessage(),
 			enrollmentService.enrollLectureByNumber(studentPk, lectureNumber));
 	}
